Rx Delay in calib stream. HWInterface update xml/dll files.

+++ @HWinterface/HWinterface.m
@@ -164,8 +164,8 @@ classdef HWinterface <handle
             if(length(m)~=1)
                 error('can set only one register');
             end
-            regVal_=cast(uint32(regVal),m.type);
-            % regVal_=cast(regVal,m.type);
+%             regVal_=cast(uint32(regVal),m.type);
+            regVal_=cast(regVal,m.type);
             if(exist('forceUpdate','var') && forceUpdate)
                 obj.m_fw.setRegs(m.regName,regVal_,'forceupdate');
             else
